Strip DatabaseQueryService. Move sparql template logic to new class SparqlQueryGenerator.
