+++ report-api/src/main/java/no/sikt/nva/data/report/api/fetch/FetchNviInstitutionReport.java
@@ -17,6 +17,7 @@ import no.sikt.nva.data.report.api.fetch.formatter.ExcelFormatter;
 import no.sikt.nva.data.report.api.fetch.formatter.PlainTextFormatter;
 import no.sikt.nva.data.report.api.fetch.model.ReportFormat;
 import no.sikt.nva.data.report.api.fetch.service.DatabaseQueryService;
+import no.sikt.nva.data.report.api.fetch.service.SparqlQueryGenerator;
 import nva.commons.apigateway.AccessRight;
 import nva.commons.apigateway.ApiGatewayHandler;
 import nva.commons.apigateway.RequestInfo;
@@ -89,7 +90,8 @@ public class FetchNviInstitutionReport extends ApiGatewayHandler<Void, String> {
     private String getResult(String reportingYear, String topLevelOrganization, ReportFormat reportFormat) {
         var replacementStrings = Map.of(REPLACE_REPORTING_YEAR, reportingYear,
                                         REPLACE_TOP_LEVEL_ORG, topLevelOrganization);
-        return queryService.getResult(NVI_INSTITUTION_SPARQL, replacementStrings, getFormatter(reportFormat));
+        var sparqlQuery = SparqlQueryGenerator.getSparqlQuery(NVI_INSTITUTION_SPARQL, replacementStrings);
+        return queryService.getResult(sparqlQuery, getFormatter(reportFormat));
     }
 
     private void validateAccessRights(RequestInfo requestInfo) throws UnauthorizedException {
