+++ report-api/src/main/java/no/sikt/nva/data/report/api/fetch/service/QueryService.java
@@ -1,9 +1,8 @@
 package no.sikt.nva.data.report.api.fetch.service;
 
 import commons.formatter.ResponseFormatter;
-import java.util.Map;
 
 public interface QueryService {
-    String getResult(String sparqlTemplate, Map<String, String> replacementStrings,
-                     ResponseFormatter formatter);
+
+    String getResult(String sparqlQuery, ResponseFormatter formatter);
 }
